Uploaded ISLA libraries
These are the libraries for ISLA to function...

EvaRes - contains the library for which will be scanned through for matching results. If the user input matches any of the preset sentences, ISLA should be able to check the category in which the user input falls and make a response accordingly. 

EzRes - contains the responses that ISLA can choose from when answering questions/statements that fall within the "easy to answer" category.

EvaEmo - Library of words with an emotion meaning attached to them. 

EmoAna - Words with basic and manageable emotion values attached to them. This library is in place to further sort out the output after running through EvaEmo to output a condense emotion analysis.